Change data type machine age from int to decimal

+++ Modules/Appraisal/Appraisal/Application/Features/Appraisals/CreateMachineryProperty/CreateMachineryPropertyCommand.cs
@@ -45,7 +45,7 @@ public record CreateMachineryPropertyCommand(
     string? Location = null,
     string? ConditionUse = null,
     string? MachineCondition = null,
-    int? MachineAge = null,
+    decimal? MachineAge = null,
     string? MachineEfficiency = null,
     string? MachineTechnology = null,
     string? UsagePurpose = null,
