+++ Modules/Appraisal/Appraisal/Application/Features/Appraisals/UpdateMachineryProperty/UpdateMachineryPropertyCommand.cs
@@ -43,7 +43,7 @@ public record UpdateMachineryPropertyCommand(
     string? Location = null,
     string? ConditionUse = null,
     string? MachineCondition = null,
-    int? MachineAge = null,
+    decimal? MachineAge = null,
     string? MachineEfficiency = null,
     string? MachineTechnology = null,
     string? UsagePurpose = null,
